ccon: Setup uid_map, setgroups, and gid_map
I handle this in the parent process, because without a new mount
namespace the child won't mount a local /proc and it doesn't know it's
PID in the host namespace to find the right /proc/{PID}/uid_map,
etc. entries there.  To avoid the child acting before it has been
mapped to root (in cases where that's the configured mapping), I've
added a new USER_NAMESPACE_MAPPING_COMPLETE message, and block the
child until the parent sends it that message.

You don't need to be root to do this.  From user_namespaces(7) [1]:

  When a user namespace is created, the kernel records the effective
  user ID of the creating process as being the "owner" of the
  namespace.  A process that resides in the parent of the user
  namespace and whose effective user ID matches the owner of the
  namespace has all capabilities in the namespace.

however, there are restrictions on what users without CAP_SETUID (or
CAP_SETGID) can do (details in user_namespaces(7) [1]).  In
particular [1]:

  In the case of gid_map, use of the setgroups(2) system call must
  first be denied by writing "deny" to the /proc/[pid]/setgroups file
  (see below) before writing to gid_map.

which is why we write setgroups before gid_map.

For a tutorial on user namepaces, see [2].

[1]: http://man7.org/linux/man-pages/man7/user_namespaces.7.html#DESCRIPTION
[2]: https://lwn.net/Articles/532593/