ccon: Add mount handling
This is a pretty thin wrapper around mount(2) [1].  I've extracted the
MS_* names from the current master snapshot of that page [2], and it
was last touched significantly in cdd14e51 (mount.2: ffix,
2015-03-03).  Command to extract all mentioned constants:

  $ curl 'http://git.kernel.org/cgit/docs/man-pages/man-pages.git/plain/man2/mount.2?id=95655a22795948470e8f2228360d8c57b8f66000' |
  >   tr '[:space:]' '\n' |
  >   sed 's/,//' |
  >   grep ^MS_ |
  >   sort |
  >   uniq

Although a number of the constants need more documentation (search [2]
for "FIXME").

I've wrapped the constants that my glibc 2.20 doesn't define in #ifdef
directives, but this whole else-if business should probably be
replaced with a macro to handle that sort of thing more uniformly.

[1]: http://man7.org/linux/man-pages/man2/mount.2.html
[2]: http://git.kernel.org/cgit/docs/man-pages/man-pages.git/tree/man2/mount.2?id=95655a22795948470e8f2228360d8c57b8f66000