ccon: Add support for a 'pivot-root' mount type
Folks who enter a mount namespace will most likely do something like:

1. Bind-mount host filesystems into their bundle's root filesystem.
2. pivot_root(2) [1] to make their bundle's root filesystem the actual
   root.
3. Perform additional setup tasks (e.g. mounting /proc, /sys,
   /dev/pts, ...).

pivot_root(2) is a separate syscall from mount(2), but by handling it
in handle_mounts we make it easy for the user to understand where it
fits in.  There are more notes on using pivot_root (including the cd
and unmount suggestions), see pivot_root(8) [2] and this article [3].
If you want to pivot onto something that's not already a mount point,
you can use the bind-mounting trick mentioned in switch_root(8) [4]:

  $ mount --bind $DIR $DIR

As mentioned in pivot_root(2), there aren't glibc bindings for this
syscall, so we need to call it directly.  See syscall(2) [5] and
syscalls(2) [6] for notes on that.

The lazy umount(2) [7] lets us disconnect the old root from the
filesystem even if we still have open references to it (e.g. the
standard streams are connected to /dev/pts/...).

[1]: http://man7.org/linux/man-pages/man2/pivot_root.2.html
[2]: http://man7.org/linux/man-pages/man8/pivot_root.8.html
[3]: http://www.ibm.com/developerworks/library/l-mount-namespaces/index.html
[4]: http://man7.org/linux/man-pages/man8/switch_root.8.html#NOTES
[5]: http://man7.org/linux/man-pages/man2/pivot_root.2.html
[6]: http://man7.org/linux/man-pages/man2/syscalls.2.html
[7]: http://man7.org/linux/man-pages/man2/umount.2.html