config.json: Add an example config
Assuming they have UID 1000 and GID 1000, users should be able to run
ccon on this config without special privileges.  It creates new
namespaces of each type, so the host kernel will need those all
compiled in.  And it will execute a BusyBox shell, so the user will
need BusyBox installed.  In the event that one of these conditions is
not satisfied, figuring out what to update should be fairly
straightforward.