ccon: Print all hook arguments (not just the path)
The logs for examples/good/net-veth-root are a lot easier to debug if
you don't have to count through the config to figure out which command
the third 'sh' hook is executing.  I'm not bothering with string
quoting or anything like that, so you can't paste these into a shell.
But they should give folks enough context in most cases.