ccon: Add process.host to fexecve from the host mount namespace
Alexander Larsson asked for a custom init process on
dev@opencontainers.org [1].  This commit adds that with fexecve to
execute a binary that may no longer be accessible from the container
mount namespace.

In the event of a new mount namespace, we'll actually be performing
the lookup in the container mount namespace to avoid having to inherit
the file descriptor through the clone and close it on the host
process.  However, the lookup is performed before entering preexisting
namespaces or performing container mounts, and at that point the
container mount namespace looks just like the host mount namespace.

[1]: https://groups.google.com/a/opencontainers.org/forum/#!topic/dev/wiRqu5oQ6y4
     Message-Id: <254f5101-5d1a-41dc-89f6-dbc9e7a59389@opencontainers.org>
     Subject: desktop sandboxes in the opencontainer initiative?