ccon: Cast write() return to size_t before comparing with strlen
Fail if the value is negative, or if the non-negative value (cast to
size_t) isn't the length we expected to write.

We could probably get away without the ssize_t n if we just cast the
strlen to ssize_t before comparing with the return value, but that
could conceivably result in overflows with crazy-long messages.  I
doubt we'd ever have messages longer than ssize_t's max, but it's easy
enough to play it safe.

Caught by Splint 3.1.2.