ccon: Add a die() handler and listen for termination signals
In the container's default block_forever, so folks don't have to
SIGKILL it.  The "termination signal" set comes from the "Term" action
in the POSIX.1-1990 table in signal(7), but we don't need a handler
for SIGKILL, we don't expect SIGPIPE or SIGALRM while blocking, and we
don't need to handle SIGUSR1 and SIGUSR2.