ccon: Add parent signal handlers for HUP, INT, and TERM
Forward the signals to the child process (inside the container), and
continue as usual.  If the container process is similar to our
block_forever, these signals will cause it to shutdown cleanly, after
which the host process will run the usual post-stop hooks and also
exit cleanly.

I've also stubbed in a reaper to handle any SIGCHLD the host gets
before the waitid call.  The idea here is that if we notice the
container-process dying for any reason, we can skip the rest of the
pre-waitid setup and get right to reaping, post-stop hooks, and other
cleanup.

Now that we have a SIGCHLD handler, we need a loop around _wait's
waitid to catch EINTR and reset waitid (more details in wait(2)).