ccon: Replace most fprintf(stderr, ...) with LOG(...)
This will make it easier for us to turn the debug logging off (by
having the macro use a flag).  I haven't gone through and separated
out debug logging from "we're bailing out, tell the caller why" (which
we might not want to put behind a --verbose flag), but we can do that
later on if it seems useful.