ccon-oci: Add a wrapper for formal OCI compliance
This way I can develop a ccon-native config that makes sense to me,
while still supporting everyone else that uses the OCI-specified
config and bundle format.  The current OCI support is not exhaustive.
For example, config.json's hostname is silently ignored, and I crash
if runtime.json says anything about cgroups.  But it handles the 0.1.0
example I've added reasonably well, and doesn't require you to be root
to run it.

The output ccon spec is for ccon v0.1.0, because no current OCI spec
requires the process.host setting added in v0.2.0.