ccon-oci: Stub in initial support for hooks
ccon doesn't believe in container IDs, so this still needs some work
before it will feed OCI-specified state JSON into the hook's stdin.

The hook syntax (with argv[0] in args) doesn't quite match the v0.1.0
docs [1,2], but it does match the phrasing from
opencontainers/specs#255 which was claimed to "clarify, but not
change" the semantics [3,4].  I think that the v0.1.0 phrasing wasn't
clear enough to decide if #255 was a change or not, but since it
wasn't clear, there's no more harm in just going with Qiang and
Mrunal's interpretation.

[1]: https://github.com/opencontainers/specs/blob/v0.1.1/runtime.md#post-stop
[2]: https://groups.google.com/a/opencontainers.org/forum/#!topic/dev/luaaorsya10
     Subject: Explicit args[0] for hooks?
     Date: Sat, 15 Aug 2015 12:15:23 -0700
     Message-ID: <20150815191523.GK21585@odin.tremily.us>
[3]: https://github.com/opencontainers/specs/pull/255#issuecomment-159786325
[4]: https://github.com/opencontainers/specs/pull/255#issuecomment-162080129