ccon: Add '--config-string JSON'
This makes some invocations easier (because we don't need to require
Bash, and it's one less layer of nesting to unpack).  It's not all
that useful if you don't already have the full JSON though
(e.g. you're missing a PID that goes in the JSON somewhere), since
you'd have to tweak the serialized bytes directly.  It does work well
if you already have the full JSON string.