README: Point to the OCI example for details on ccon-oci
It would be distracting to talk about that wrapper in detail here, but
folks interested in the wrapper will want more details.  Linking to
the ccon-oci example and putting the discussion there gives us more
space for details.  After we add support for additional OCI versions,
we'll probably want to move the generic ccon-oci description to
examples/good/oci/README.md or some such.