ccon: Add a single Jansson schema validation pass
Jansson docs are at [1].  Their format language does not support
language for "all entries in this array must support the following
schema", so I've left array-value checking out for now.  And the
format string (and associated array of key strings) is awkward enough
that I didn't want repeated strings for v0.2.0 and v0.1.0 (with and
without process.host), so ensuring a v0.1.0 config doesn't contain
process.host is left to a follow-up commit.  To keep the format string
and associated key strings somewhat readable, I've disabled indent
for that region [2] and setup some nested indents.

The FIXME comment works around a bug (at least Gentoo's
dev-util/indent-2.2.11-r1).

  $ cat <<EOF >test.c
  int main()
  {
  	if (1) {
  		return 0;
  	}

  /* *INDENT-OFF* */
  	return 1;
  /* *INDENT-ON* */

  /*
  	return 0;
  }
  EOF
  $ indent --ignore-profile --linux-style text.c
    (this removes the blank line before *INDENT-OFF*)
  $ indent --ignore-profile --linux-style text.c
  indent: x.c:12: Error:Unexpected end of file
    (this injects a comment-opener "/*" on the line after *INDENT-ON*)

Jansson's errors aren't particularly useful (as of v2.7).  For
example:

  $ ccon --verbose --config-string '{"version": "0.1.0", "a": 1}'
  validation error: 1 object item(s) left unpacked
  config.json invalid

which could at least mention that the "left unpacked" object as key
"a".  But that's something that needs a Jansson patch, not a ccon
workaround [3].  I've skipped error.line and error.column in this
error, because they reference the format string, not the user's config
JSON.  Jansson doesn't seem to preserve a JSON object's source
location after parsing the config file.

[1]: http://jansson.readthedocs.org/en/2.7/apiref.html#parsing-and-validating-values
[2]: https://www.gnu.org/software/indent/manual/html_node/indent_13.html#SEC13
[3]: https://github.com/akheron/jansson/pull/263