ccon: Set parent death signal to SIGKILL
In 193e08db (ccon: Add parent signal handlers for HUP, INT, and TERM,
2015-10-22) the host process started catching some signals, sending
KILL to the container process and then continuing.  However, we *also*
want the container process to be killed when the host process dies
(e.g. if the host process is KILLed).  Using prctl(2) to set the
parent death signal lets us do this.

With this extra safety net, we might want to adjust kill_child to
forward the signal it catches (so a HUP sent to the host process
results in a HUP sent to the container process, instead of a KILL sent
to the container process).  I've left that for a subsequent commit (if
we do it at all).