test/t1001-process-user.t: Test process.user
Both with and without privileges (without fails) and also for hooks,
using id(1) [1].  The unprivileged calls fail in the absence of a user
namespace; we'll ensure they work there when we add user-namespace
tests.

Also bump Sharness to its current tip to pull in sharness.d [2], which
gives us a convenient place to put common tests like "Are we root?".
The 'true' in root.sh lets us source the script without an error
regardless of whether we're root or not.

posix-utilities.sh centralizes tests for various POSIX utilities [3].
I imagine most systems with the POSIX shell required by Sharness will
have these utilities, but it's easy enough to check, and skipping a
test because a required utility is missing will be less confusing than
failing the test because of a non-ccon issue.

[1]: http://pubs.opengroup.org/onlinepubs/9699919799/utilities/id.html
[2]: https://github.com/mlafeldt/sharness/pull/33
     529cad5 (Merge pull request #33 from grondo/project-integration,
     2015-09-06)
[3]: http://pubs.opengroup.org/onlinepubs/9699919799/idx/utilities.html