README: Remove path/other incompatibility for namespace objects
This was initially based on OCI's limitation [1].  But I never felt
that strongly about that limitation [2,3].  And join-and-tweak is
occasionally useful (see namespaces.mount in
examples/good/create-exec-delete-root/exec.json).  We never enforced
the nominal restriction, so this commit updates the docs to allow
join-and-tweak.  Users who try to tweak nonsensical things will get
errors from the kernel rejecting the invalid tweaks.

[1]: https://github.com/opencontainers/specs/pull/158
[2]: https://github.com/opencontainers/specs/pull/158#issuecomment-138687129
[3]: https://github.com/opencontainers/specs/pull/158#issuecomment-138997548