ccon-oci: Add support for process.terminal
Now that I understand what it's about and ccon supports it.

I updated the instructions to echo into ccon-oci's stdin instead of
attempting to drive it from the calling terminal.  As the README docs
for process.terminal explain, when terminal is set true, we want raw
keystroke-level stuff going in on stdin and get raw video-level stuff
out on stdout.  Instead of trying to do that from a cooked terminal,
just pipe the keystrokes in on stdout.  If you already have a terminal
and want an interactive OCI container, you should *unset*
process.terminal so the container process inherits the same terminal
standard streams that any process invoked from your terminal would
get.