test/t2002-mount-namespace.t: Test mount namespace creation
The namespaces(7) example uses readlink to figure out which namespace
a /proc/{pid}/ns/{namespace} file reference.  You could also use
stat(1), although neither are in POSIX 2013 [1].  Somewhat
surprisingly for me, the inode that's part of the link target text is
the inode of the link target:

  $ readlink /proc/self/ns/mnt
  mnt:[4026531840]
  $ stat --format '%i' /proc/self/ns/mnt
  2150346
  $ stat --dereference --format '%i' /proc/self/ns/mnt
  4026531840

The remount is the sort of thing you'd do to make a portion of your
container filesystem read-only.  There's some historical background on
the "bind and remount" approach in [2], and a remount,ro,bind example
in mount(8).

The touch stderr and exit code for a read-only filesystem aren't
defined in POSIX [3].  The read-only test passes with GNU coreutils
8.23, but I'm not sure how portable it will be.  If it turns out to be
unreliable, we'll probably have to parse /proc/self/mounts with sed
(proc(5) points to fstab(5) as format documentation) and look for a
'ro' option.

[1]: http://pubs.opengroup.org/onlinepubs/9699919799/idx/utilities.html
[2]: http://lwn.net/Articles/281157/
[3]: http://pubs.opengroup.org/onlinepubs/9699919799/utilities/touch.html