ccon: Better error handling in handle_parent
Previously socket trouble could lead to the parent process exiting
after running the pre-start hooks but without running the post-stop
hooks.  Similarly, pre-start failures would set 'err' but that value
was never used.  This commit cleans up the handling, so all
post-pre-start failures set err, try and kill the child process, and
jump to 'wait'.  After the container process is reaped into 'exit' and
the post-stop hooks are run, we return 'err' (if there was an error in
handle_parent) or 'exit' (if the handle_parent code went off
smoothly).