ccon-oci: Add 'delete' and 'event'
Delete is currently used as an internal, post-stop hook in 'create',
but you could call it directly if the 'create' process dies before
running the hook.

Event is a wrapper around the inotify-pubsub commands [1].  The
listener will currently continue listening after container deletion in
--monitor mode (or if it hasn't seen a matching event yet).  Callers
who care about that should watch for 'stopped' and kill the watcher by
hand.

[1]: https://github.com/wking/inotify-pubsub