ccon: Add --pseudoterminal-socket-path
Passing the pseudoterminal master out allows the host process to exit
early.  If you don't pass the master out, the host process has to hang
around to splice its standard streams to the master.  An early-exit
host process makes collecting the container process's exit code more
complicated (via PR_SET_CHILD_SUBREAPER or ptrace or some such), but
that may be worth the trouble for some callers.

This commit does not provide an early-exit option, but the
--pseudoterminal-socket-path option allows folks to test master
pass-out to see how that would work.