ccon: Set PATH from process.env for non-host execution
Because execvpe is using current PATH, not the one we're passing in
via its argument.  The exec_fd guards don't have a semantic impact,
because we've already opened the file descriptor using the original
PATH; they're just a performance optimization, because there's no need
to iterate through env or call setenv for the host-executable case.