ccon: Use original stderr for logging after set_terminal
By dup()ing the original stderr to some other file descriptor and
closing that descriptor just before executing another program.  This
avoids polluting the terminal output with ccon logging even if ccon is
called with --verbose.

I've also set an explicit initial value for 'verbose'.  The lack of
one didn't seem to be causing problems, but it doesn't hurt to be
explicit ;).