ccon: Shift pre-start hooks to post-create
Now that create and start are separate, the end-of-create hooks are
more appropriately named "post-create".

This commit parses the version major/minor into a float with strtof
[1].  We can ignore the patch level, because that doesn't impact the
config structure.  For systems where the radix is not a period, we use
newlocale [2], uselocale [3], and freelocale [4] to set the POSIX
LC_NUMERIC locale [5] before calling strtof and then reset the locale
afterward.  To distinguish between configs before and after the
rename, I've nominally bumped the version to 0.5.0.  I'll try and cut
a release soon to avoid dangling that change for too long.

I haven't poked at ccon-ced recently, but I'm bumping it along here
anyway.  I've dropped it's 'host: true' to catch up with 6224b0ce
(README: Document lack of 'host' for hook processes, 2015-12-18).

[1]: http://pubs.opengroup.org/onlinepubs/9699919799/functions/strtof.html
[2]: http://pubs.opengroup.org/onlinepubs/9699919799/functions/newlocale.html
[3]: http://pubs.opengroup.org/onlinepubs/9699919799/functions/uselocale.html
[4]: http://pubs.opengroup.org/onlinepubs/9699919799/functions/freelocale.html
[5]: http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap07.html#tag_07_03_04_01