ccon: Install signal handlers before cloning the child
From clone(2) [1]:

> Unlike fork(2), clone() allows the child process to share parts of
> its execution context with the calling process, such as the virtual
> address space, the table of file descriptors, and the table of
> signal handlers.
> ...
> If CLONE_SIGHAND is not set, the child process inherits a copy of
> the signal handlers of the calling process at the time clone() is
> called.

By setting the handlers before the clone call, we know they're in
place to catch signals sent immediately after the clone.  By having
clone write to child_pid and cpid, we ensure that that information is
available to the signal handlers as soon as possible (child_pid) and
that run_container knows what clone returned regardless of intervening
handler action (cpid).

It's not a big deal to have these handlers in place in the child
process, because child_pid and hook_pid will always be -1 there (and
I've explicitly initialized the static variables to make that more
obvious).  I've reset them for SIGTERM, SIGINT, and SIGHUP to get
more-efficient child termination (the default action) in those cases.

I've left the SIGCHLD handler on the child process so that it can reap
any orphan container processes if it happens to be PID 1 in the
container namespace.

I'm also checking for sigemptyset errors and fixing 'signal' ->
'sigaction' PERROR values.

[1]: http://man7.org/linux/man-pages/man2/clone.2.html