ccon: Install signal handlers before cloning the child
Block signals around clones and forks and uninstall SIGHUP, SIGINT,
and SIGTERM forwarders in the hooks.  Blocking the signals is a better
cover for the "what if we get this signal right after the fork/clone"
than the approach taken in 8f2a057 ccon: Install signal handlers
before cloning the child, 2018-03-02).  The old approach had two
problems:

* Before the container clone, a SIGTERM (say) received by the parent
  process would be silently ignored because there was not yet a
  container process to forward it to.  Now we don't install the
  handlers until we have blocked the signals.

* After the container clone, a SIGTERM (say) hitting the parent would
  be forward (appropriately) to the child, but the child would
  (inappropriately) catch it with a no-op kill_children run.  Now the
  child has uninstalled that handler by the time it unblocks the
  signal, so it will terminate when the parent sends it the SIGTERM.

There's no real need to install the handlers before the clone; they
could be installed afterwards in the parent before unblocking the
signals.  I've installed them before the clone to mirror the
hook-forking approach (where the handlers will already be installed
before the fork).

This commit adds helpers to make this easier to follow.  I'm not
terribly excited about the asymmetry between install_signal_handlers
(which sets a SIGCHLD handler) and uninstall_signal_handlers (which
doesn't unset the SIGCHLD handler).  But
uninstall_hup_int_term_handlers seemed too wordy.  And we want to keep
the SIGCHLD handler for the container process, so we continue to reap
children if that process is the PID 1 in a new continer namespace.  It
doesn't matter either way for the hook processes; they're unlikely to
receive SIGCHLDs in the short time before the hook exec call, and if
they do reaping the child is fine, and the reaped child is not going
to match child_pid or hook_pid.

The additional logging helps my sanity, but the t0001 changes could
race between the parent and child unblocks.  In my local testing the
ordering seems stable enough, so I haven't done anything about it yet.
If it becomes a problem, we can use grep to remove the unblock lines.